% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dymium-package.R
\docType{package}
\name{dymium-package}
\alias{dymium}
\alias{dymium-package}
\title{dymium: Easy Install and Load Stable Dymium Packages}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

The package makes it convenient to install and load all of the stable Dymium 
  packages. This currently includes just the [dymiumCore](www.github.com/dymium-org/dymiumCore package).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dymium.org}
  \item \url{https://github.com/dymium-org/dymium}
  \item Report bugs at \url{https://github.com/dymium-org/dymium/issues}
}

}
\author{
\strong{Maintainer}: Amarin Siripanich \email{amarin@dymium.org}

Authors:
\itemize{
  \item Taha Rashidi
}

}
\keyword{internal}
