% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{dymium_deps}
\alias{dymium_deps}
\title{List all dymium dependencies}
\source{
tidyverse
}
\usage{
dymium_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
dymium packages.}
}
\description{
List all dymium dependencies
}
