% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{dymium_update}
\alias{dymium_update}
\title{Update dymium packages}
\source{
tidyverse
}
\usage{
dymium_update(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also check all dependencies of
dymium packages.}
}
\description{
This will check to see if all dymium packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
dymium_update()
}
}
